refactor: Controller 분리 및 장소 추가 요청 변경
