+++ src/main/java/com/goormcoder/ieum/controller/PlanController.java
@@ -1,14 +1,9 @@
 package com.goormcoder.ieum.controller;
 
 
-import com.goormcoder.ieum.dto.request.PlaceCreateDto;
-import com.goormcoder.ieum.dto.request.PlaceShareDto;
 import com.goormcoder.ieum.dto.request.PlanCreateDto;
 import com.goormcoder.ieum.dto.response.DestinationFindDto;
-import com.goormcoder.ieum.dto.response.PlaceFindDto;
-import com.goormcoder.ieum.dto.response.PlaceInfoDto;
 import com.goormcoder.ieum.dto.response.PlanInfoDto;
-import com.goormcoder.ieum.service.PlaceService;
 import com.goormcoder.ieum.service.PlanService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -16,9 +11,6 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,9 +24,6 @@ import java.util.UUID;
 public class PlanController {
 
     private final PlanService planService;
-    private final PlaceService placeService;
-
-    private final SimpMessageSendingOperations messagingTemplate;
 
     @GetMapping
     @Operation(summary = "여행지 목록 조회", description = "여행지 목록을 조회합니다.")
@@ -47,20 +36,6 @@ public class PlanController {
     public ResponseEntity<PlanInfoDto> createPlan(@Valid @RequestBody PlanCreateDto planCreateDto) {
         UUID memberId = getMemberId();
         return ResponseEntity.status(HttpStatus.OK).body(planService.createPlan(planCreateDto, memberId));
-    }
-
-    @PostMapping("/pre-place")
-    @Operation(summary = "장소 추가", description = "사용자별로 방문하고 싶은 장소를 추가합니다. 카테고리 유형 - 1(명소) 또는 2(식당/카페) 또는 3(숙소)")
-    public ResponseEntity<PlaceInfoDto> createPlace(@RequestBody PlaceCreateDto placeCreateDto) {
-        UUID memberId = getMemberId();
-        return ResponseEntity.status(HttpStatus.OK).body(placeService.createPlace(memberId, placeCreateDto));
-    }
-
-    @MessageMapping("/share-place")
-    public void addPlace(@Payload PlaceShareDto placeShareDto) {
-        // UUID memberId = getMemberId(); - 검증 추가 예정
-        PlaceFindDto placeFindDto = placeService.sharePlace(placeShareDto);
-        messagingTemplate.convertAndSend("/sub/plans/" + placeShareDto.planId(), placeFindDto);
     }
 
     private UUID getMemberId() {
