+++ src/main/java/com/goormcoder/ieum/dto/request/PlaceCreateDto.java
@@ -2,7 +2,6 @@ package com.goormcoder.ieum.dto.request;
 
 public record PlaceCreateDto(
 
-        Long planId,
         String placeName,
         String address,
         Long categoryId
